Factor out pagination into reusable method.
